backend/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
